feat(view): 선택된 사용자 로직 컴포넌트로 분리 (#381)
* feat(view): FilteredAuthors 컴포넌트 추가

* feat(view): TempoeralFilter 컴포넌트에서 FilteredAuthors code 분리

* feat(view): FilteredAuthors 컴포넌트 App 컴포넌트에 위치
