+++ packages/view/src/components/TemporalFilter/TemporalFilter.scss
@@ -102,13 +102,3 @@
     color: black;
   }
 }
-
-.selected-cluster {
-  position: relative;
-  top: 30px;
-  display: flex;
-  flex-wrap: wrap;
-  width: 100%;
-  padding: 4px 6px;
-  box-sizing: border-box;
-}
