+++ packages/view/src/components/index.ts
@@ -4,3 +4,4 @@ export * from "./TemporalFilter";
 export * from "./ThemeSelector";
 export * from "./VerticalClusterList";
 export * from "./BranchSelector";
+export * from "./FilteredAuthors";
